==> postprocessing/01_import_data_make_adata.ipynb <==
import scanpy as sc
import os,sys
import datetime
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import re
import scipy

sc.logging.print_header()

## Importing data 

#function of current time to add to filenames
def now():
    return datetime.datetime.now().strftime('%y%m%d_%Hh%M')

#a list of paths where the data is stored. CD34+ data is used as an example.
path_list =[
'/Users/karolisgoda/Desktop/cluster/CD34/CD34-04/GeneFull/raw',
'/Users/karolisgoda/Desktop/cluster/CD34/CD34-01/GeneFull/raw',
'/Users/karolisgoda/Desktop/cluster/CD34/CD34-02/GeneFull/raw',
'/Users/karolisgoda/Desktop/cluster/CD34/CD34-05/GeneFull/raw',
'/Users/karolisgoda/Desktop/cluster/CD34/CD34-03/GeneFull/raw'
]


#backup directory
!mkdir -p backups

#making dict of adatas where key is library name and value is adata object
#this is not necessary and could be replaced with a list

dic_ad = {}

#loading data
for path in path_list:
    
    #force slach at the end
    if path[-1] != '/':
        path = path + '/'
    
    #extract_library name
    parsed = path.split('/')
    name = parsed[-4]

    # load to AnnData
    adata = sc.read(path  + 'matrix.mtx', cache=False)
    adata = adata.T
    adata.var_names = pd.read_csv(path + 'features.tsv', header = None, sep='\t' )[1]
    adata.obs_names = pd.read_csv(path + 'barcodes.tsv',  header = None)[0]
    adata.var_names_make_unique()
    
    # add library to obs
    adata.obs['library'] = name
    
    #save unfiltered adata objects
    fname1 = 'backups/raw_%dx%d_%s_%s.h5ad'%(adata.shape[0],adata.shape[1], name, now())
    print(fname1)
    #adata.write(fname1)
    
    
    # add to dict
    print(name, adata.shape)
    dic_ad[name] = adata
            

#create list of anndatas
lad=list(dic_ad.values()) 

lad[0]

# gene names must be the same for all libraries, but double-check
v0 = lad[0].var_names
print(np.prod([np.prod(v0==i.var_names) for i in lad]).astype(bool))

# also check that gene names are unique:
print(len(v0)==len(set(v0)))


#concatenate manually, since sc.AnnData.concatenate() can only take 14 arguments

Xcat = scipy.sparse.vstack([i.X for i in lad])
obscat = pd.concat([i.obs for i in lad])
varnames = lad[0].var_names


adata = sc.AnnData(Xcat) 
adata.obs = obscat
adata.var_names = varnames

# index
adata.obs_names = np.arange(adata.shape[0]).astype(str)


print(adata.shape)

#save the huge file for later pre-processing
fname = 'backups/concatenated_raw_%dx%d_%s.h5ad'%(adata.shape[0],adata.shape[1], now())
print(fname)
adata.write(fname)
