==> postprocessing/03_adding_technical_information.ipynb <==
import scanpy as sc
import os,sys
import datetime
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import math
import matplotlib.ticker as mticker

# add the utility function folder to PATH
sys.path.append(os.path.abspath("utility_functions/"))

from rz_import_statements import *
import rz_functions as rz
import rz_utility_spring as srz

# Load  data

adata = sc.read_h5ad('backups/filt_raw_20021x37733_230928_14h17.h5ad')


## Adding some technical information to data

#adding info about cell preservation
for i in sorted(adata.obs['conditions2'].unique()):
    print('"%s":"",'%i)

renamer = {
"10X fixed":"Fixed",
"10X fresh":"Fresh",
"InDrop2 fixed":"Fixed",
"InDrop2 fresh":"Fresh",
}

adata.obs['preservation'] = [renamer[i] for i in adata.obs['conditions2']]
adata.obs['preservation']

#adding info about method
renamer = {
"10X fixed":"10X",
"10X fresh":"10X",
"InDrop2 fixed":"InDrop2",
"InDrop2 fresh":"InDrop2",
}

adata.obs['method'] = [renamer[i] for i in adata.obs['conditions2']]
adata.obs['method']

## Saving the updated obs file

fname = 'backups/obs_info_technical_%dx%d_%s.npz'%(adata.obs.shape[0],adata.obs.shape[1],rz.now())
print(fname)
rz.save_df(adata.obs,fname)
