==> postprocessing/12_cluster_graph.ipynb <==
import scanpy as sc
import scanpy.external as sce
#sc.logging.print_versions()
#sc.logging.print_memory_usage()
#sc.settings.verbosity = 2
import os,sys
import datetime
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import math
import matplotlib.ticker as mticker

# add the utility function folder to PATH
import sys
sys.path.append(os.path.abspath("utility_functions/"))

from rz_import_statements import *
import rz_functions as rz
import rz_fig_params # this adjust mpl.rcParams, almost nothing to import, import after scanpy to overwrite rc.Params
import rz_utility_spring as srz

## Load data

#load the newest obs with doublet and RBC removal
filename = 'backups/obs_info_clean_good_15372x30_231009_12h06.npz'
encoding = 'latin1'

with np.load(filename,encoding=encoding, allow_pickle = True) as f:
    obs = pd.DataFrame(**f)

# load intermediates I saved when preparing the SPRING plot
path1 = "/Users/karolisgoda/Library/CloudStorage/GoogleDrive-karolisgoda@gmail.com/My Drive/MF/data/SPRING_dev-master/data/" 
project_dir = path1+'CD34_good/clean_good_sc_hvg/'
plot_name = 'palantir_cell_cycle'


#params = rz.load_stuff(project_dir+plot_name+'/params.pickle')
#params.keys()

#the umap and graph data is saved in adata file
adatag = sc.read_h5ad('backups/clean_good_sc_hvg_15372x3324_231129_15h00.h5ad')


adatag

adatag.X = adatag.layers['X_lin_cptt']

G  = adatag.obsp['connectivities'] #selecting the graph

adatag.X.sum(axis=1)

# Leiden

cg = {}
for i in [0.2, 0.4, 0.6, 0.8, 1]:
    mock = adatag.copy()
    sc.tl.leiden(mock, resolution=i)
    l = list(mock.obs['leiden'])
    cg['leiden_res_%.1f'%i] = l
    print(len(set(l)))

# Phenograph clustering

# Clustering using PhenoGraph
for i in [0.1, 0.2, 0.3, 0.4]:
    sc.external.tl.phenograph(adatag, clustering_algo='leiden', k=15, jaccard=True, primary_metric='euclidean', 
                          resolution_parameter = i)
    l = list(adatag.obs['pheno_leiden'])
    cg['pheno_leiden_res_%.1f'%i] = l
    print(len(set(l)))

# Spectral clustering (choose number of clusters)

# this one is much slower
for i in [
    9,10,11,12,13]:
    key = 'sp_cl_%d'%(i)
    print(key)
    cg[key] = list(srz.spec_clust(G,i).astype(str))
    print(key)

# Append result to the same SPRING plot

# load current color dictionary
cg0 = srz.read_cell_groupings(project_dir+plot_name+'/categorical_coloring_data.json')


# color dictionary of dictionaries
cdd = {key:value['label_colors'] for key,value in cg0.items()}


cg = {key:list(np.array(value).astype(str)) for key,value in cg.items()}

project_dir+plot_name+'/categorical_coloring_data.json'

# append categorical colortrack
srz.append_cell_groupings(project_dir+plot_name,cg,colordd=cdd)

for key, value in cg.items():
    adatag.obs[key] = cg[key]

obs1 = adatag.obs

#save obs with clusters

fname = 'backups/obs_info_sc_hvg_cell_cycle_cluster_%dx%d_%s'%(obs1.shape[0],obs1.shape[1],rz.now())
print(fname)
rz.save_df(obs1,fname)
