==> postprocessing/13_dge_global_spectral_clusters.ipynb <==
### Import scanpy

import scanpy as sc
import scanpy.external as sce
#sc.logging.print_versions()
#sc.logging.print_memory_usage()
#sc.settings.verbosity = 2
import os,sys
import datetime
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import math
import matplotlib.ticker as mticker

# add the utility function folder to PATH
sys.path.append(os.path.abspath("utility_functions/"))

from rz_import_statements import *
import rz_functions as rz
import rz_fig_params # this adjust mpl.rcParams, almost nothing to import, import after scanpy to overwrite rc.Params
import rz_utility_spring as srz

# Load data

adata = sc.read_h5ad('backups/final_sc_hvg_palantir_15372x3324_231130_12h58.h5ad') 

# overwrite obs with the most recent version
filename = 'backups/obs_info_annotated_sc_hvg_no_cycle_15372x25_231129_17h30.npz'
encoding = 'latin1'

with np.load(filename,encoding=encoding, allow_pickle = True) as f:
    obs = pd.DataFrame(**f)
adata.obs = obs

### Scale (normalize) data

adata = adata.raw.to_adata()

# turn into counts per 10k
print(adata.X[:5,:].sum(axis=1))
print()
sc.pp.normalize_per_cell(adata, counts_per_cell_after=1e4)
print(adata.X[:5,:].sum(axis=1))

# Get enriched genes

Upon interactive exploration of the UMAP in the SPRING (Weinreb et al., 2018) application, it was decided that spectral clustering with parameter k=43 is the best representation to separate the cellular phenotypes present in the dataset.

#taking all cells now

cmask = np.repeat(True,adata.shape[0])
print(cmask.sum())

#spectral clustering labels
for i in adata.obs:
    if i.startswith('cell_type'):
        print (i)

# get centroids selecting the cluster configuration with k=43
thelabel = 'cell_type'
centroids = rz.centroids(thelabel,adata[cmask])

# For each cluster, find genes that are statistically significantly higher or lower in cluster x compared to
# all other cells collectively

#label-free filter to remove low abundance genes - 
#gene has to be expressed in at least min_cells by at least min_counts
min_counts = 10
min_cells = 5

gmask = srz.filter_abund_genes(adata.X[cmask], min_counts, min_cells)

mwu_dict = {}
start=time.time()
counter=0

meta = adata[cmask].obs
E = adata[cmask].X
gene_list = adata.var_names


for cluster in meta[thelabel].unique():
    counter+=1
    mask1 = (meta[thelabel]==cluster).values
    mask2 = mask1==False
    
    cg1 = np.array(E[:,gmask][mask1,:].todense())
    cg2 = np.array(E[:,gmask][mask2,:].todense())
    mwu_dict[cluster] = rz.mwu(cg1,cg2,genes=gene_list[gmask],print_progression=True)
    print("%d/%d"%(counter,len(meta[thelabel].unique())))
    print(cluster, 'done',cg1.shape[0]+cg2.shape[0])
print(time.time()-start)

fname = 'backups/cell_types_vs_rest_MWU_result_dict_no_cycle_%s%s'%(rz.now(), thelabel)
print(fname)
rz.save_stuff(mwu_dict,fname)

# if continuing from backup
#mwu_dict = rz.load_stuff('backups_JZ_2022/cluster_vs_rest_MWU_result_dict_220317_17h32sp_cl_43.pickle')

# select pseudovalue to add
pseudo = 1 # 1 counts per 10k

thelabel = 'cell_type'

print(thelabel)

# fcdict - fold-change dictionary

fcdict = rz.get_fc_to_all_other(
        lab = thelabel,
        meta = adata[cmask].obs,
        E = adata[cmask].X,
        pseudo = pseudo,
        gene_list = adata.var_names,
        )

# leave only genes with a significant difference.
fcdictsig = {}

# before the mwu test, I prefiltered genes on abundance, apply this mask here as well
print(gmask.sum())
for key,value in fcdict.items():
    sigmask = (mwu_dict[key]['fdr']<0.05).values
    fcdictsig[key] = value[gmask][sigmask]
    print(key,sigmask.sum(),len(fcdictsig[key]))

# nr genes to consider:
upto = 100 # up to 100 genes used to generate Table S1


frame = {}
for key,value in fcdictsig.items():
    s = value.sort_values(ascending=False)[:upto]
    key2 = str(key)+'_FC'
    frame[str(key)] = s.index
    frame[key2] = s.values
frame = pd.DataFrame(frame)
frame[[i for i in frame.columns if "FC" in i]].min() #ok, all above 1.

outdir = 'outputs/'

fname = outdir+'lists_enriched_genes_no_cycle_top_%d_%s_%s.xlsx'%(upto,thelabel,rz.now())
print(fname)
frame.to_excel(fname)
